添加删除论文接口

+++ 221801310&221801325/backend/src/main/java/com/pair/papercrawler/controller/SearchController.java
@@ -84,7 +84,10 @@ public class SearchController {
 
     @DeleteMapping("/deletePaper")
     public ResponseMessage deletePaper(@RequestParam(name = "paperId") Integer paperId){
-        papersService.deletePaperByPaperId(paperId);
-        return ResponseMessage.success("successfully delete");
+        if (papersService.selectPaperByPaperId(paperId)!=null){
+            papersService.deletePaperByPaperId(paperId);
+            return ResponseMessage.success("successfully delete");
+        }
+        return ResponseMessage.failure("faliure deletw");
     }
 }
